patch: Adding existing secret option
This introduces the option for an existing secret to be supplied with the credentials for postgres, smtp, redis and s3.
By allowing users to roll their own secrets for these values, we improve their security as they don't have to hard code any values into their supplied `values.yaml` file.

The presumption is that if a user is supplying some credentials via a secret then all of the listed ones should be supplied.